Sample method overriding
In the Main class, we create an instance of the Animal class and call its makeSound() method, which outputs "Some sound" to the console. We also create instances of the Cat and Dog classes and call their makeSound() methods, which output "Meow" and "Woof" to the console, respectively. This demonstrates how the makeSound() method in the subclass overrides the implementation of the same method in the superclass.